data importer simplified demo data changed to DM-DS workshop scenario
