+++ core/management/commands/load_demo_data.py
@@ -2,7 +2,7 @@ from django.core.management import BaseCommand
 from django.conf import settings
 import os
 
-from core.importer.datadecs_importer import DatadecsImporter
+
 from core.importer.datasets_importer import DatasetsImporter
 from core.importer.projects_importer import ProjectsImporter
 from core.models import User
@@ -28,15 +28,6 @@ class Command(BaseCommand):
                 importer = DatasetsImporter()
                 importer.import_json(json_file_contents)
                 self.stdout.write(self.style.SUCCESS("Dataset import  successful!"))
-            datadecs_json = os.path.join(DEMO_DATA_DIR, 'datadecs.json')
-            with open(datadecs_json, encoding='utf-8') as json_file:
-                json_file_contents = json_file.read()
-                importer = DatadecsImporter()
-                importer.import_json(json_file_contents)
-                self.stdout.write(self.style.SUCCESS("Data declaration import  successful!"))
-
-
-
             admin_usr = User.objects.create_user(username='admin', password='', email='demo.admin@uni.lu')
             admin_usr.is_superuser =True
             admin_usr.save()
@@ -48,8 +39,6 @@ class Command(BaseCommand):
                     user.is_staff = True
                     user.set_password('demo')
                     user.save()
-
-
 
         except Exception as e:
             self.stderr.write(
