--- core/tests/data/datadecs.json
@@ -1,136 +1,0 @@
-[
-  {
-    "dataset": "ABCD data",
-    "title": "ABCD",
-    "data_type_notes": "..\n",
-    "data_types": [
-      "Metabolomics",
-      "Clinical_data",
-      "Methylation_array"
-    ],
-    "de_identification": "anonymization",
-    "subject_categories": "controls",
-    "has_special_subjects": true,
-    "special_subject_notes": "2 year old children",
-    "source_notes": "Data is from collaborator.",
-    "use_restrictions": [
-      {
-        "ga4gh_code": "PS",
-        "note": "Use is restricted to projects: ABCD"
-      },
-      {
-        "ga4gh_code": "PUB",
-        "note": "Acknowledgement required."
-      }
-    ]
-  },
-  {
-    "dataset": "Hypertension data",
-    "title": "Hypertension-ABC disease",
-    "data_types": [
-      "Genotype_data",
-      "Whole_genome_sequencing"
-    ],
-    "de_identification": "pseudonymization",
-    "subject_categories": "cases_and_controls",
-    "has_special_subjects": false,
-    "source_collaboration": {
-      "collab_inst": "ELU_I_94",
-      "collab_pi": "Alberto Pico",
-      "collab_project": "Hypertension",
-      "collab_role": "joint_controller"
-    },
-    "source_notes": "Data is from collaborator.",
-    "use_restrictions": [
-      {
-        "ga4gh_code": "PS",
-        "note": "Use is restricted to projects: Hypertension"
-      },
-      {
-        "ga4gh_code": "RS-[XX]",
-        "note": "Use is restricted to research areas: Hypertension-ABC disease"
-      }
-    ]
-  },
-  {
-    "dataset": "PD data",
-    "title": "XYZ",
-    "source_notes": "Data is from own cohort.",
-    "data_types": [
-      "Cell_Imaging",
-      "Clinical_data",
-      "Other_Phenotype_data",
-      "Samples"
-    ],
-    "de_identification": "pseudonymization",
-    "ombudsman": "Biobank, Principle Investigator",
-    "subject_categories": "cases_and_controls",
-    "has_special_subjects": false,
-    "consent_status": "homogeneous",
-    "use_restrictions": [
-      {
-        "ga4gh_code": "PS",
-        "note": "Consent form restricts data use to projects XYZ"
-      },
-      {
-        "ga4gh_code": "RS-[XX]",
-        "note": "Data is consented for research onParkinson's disease"
-      },
-      {
-        "ga4gh_code": "GS-[XX]",
-        "note": "Data is consented for sharing outside institute (Within Luxembourg)"
-      }
-    ]
-  },
-  {
-    "dataset": "REPO data",
-    "title": "ZZZZ",
-    "source_collaboration": {
-      "collab_inst": "ELU_I_84"
-    },
-    "source_notes": "Data is obtained from repository.",
-    "data_types": [
-      "Whole_genome_sequencing",
-      "Other_Phenotype_data"
-    ],
-    "de_identification": "pseudonymization",
-    "subject_categories": "cases_and_controls",
-    "has_special_subjects": false,
-    "access_category": "controlled_access",
-    "use_restrictions": [
-      {
-        "ga4gh_code": "PUB",
-        "note": "Acknowledgement required."
-      }
-    ]
-  },
-  {
-    "dataset": "PD data",
-    "title": "PD data (german cohort)",
-    "data_type_notes": "Small-scale protein, mRNA or mtDNA integrity data, cell and mitochondrial function data\n",
-    "data_types": [
-      "Other",
-      "Samples"
-    ],
-    "de_identification": "pseudonymization",
-    "subject_categories": "cases_and_controls",
-    "has_special_subjects": false,
-    "source_collaboration": {
-      "collab_inst": "ELU_I_8",
-      "collab_pi": "Manuela Swift",
-      "collab_project": "PD_Project",
-      "collab_role": "joint_controller"
-    },
-    "source_notes": "Data is from collaborator.",
-    "use_restrictions": [
-      {
-        "ga4gh_code": "PUB",
-        "note": "Acknowledgement required."
-      },
-      {
-        "ga4gh_code": "TS-[XX]",
-        "note": "Data is obtained for a limited duration. 2021-02-28"
-      }
-    ]
-  }
-]
