+++ data/demo/projects.json
@@ -1,39 +1,76 @@
+
 [
     {
-        "acronym": "DAISY-Dev",
-        "title": "Data Information System DAISY Development.",
-        "description": "Development of software to record GDPR-relevant metadata of human data and biosamples used in research.",
-        "start_date": "2018-03-01",
-        "end_date": "2019-09-01",
+        "acronym": "EPIC-DREM",
+        "name": "Understanding the gene regulatory interactions underlying cell differentiation and identity",
+        "description": "\"Temporal data on gene expression and context-specific open chromatin states can improve identification of key transcription factors (TFs) and the gene regulatory networks (GRNs) controlling cellular differentiation. However, their integration remains challenging. Here, we delineate a general approach for data-driven and unbiased identification of key TFs and dynamic GRNs, called EPIC-DREM. We generated time-series transcriptomic and epigenomic profiles during differentiation of mouse multipotent bone marrow stromal cell line (ST2) toward adipocytes and osteoblasts. Using our novel approach we constructed time-resolved GRNs for both lineages and identifed the shared TFs involved in both differentiation processes. To take an alternative approach to prioritize the identified shared regulators, we mapped dynamic super-enhancers in both lineages and associated them to target genes with correlated expression profiles. The combination of the two approaches identified aryl hydrocarbon receptor (AHR) and Glis family zinc finger 1 (GLIS1) as mesenchymal key TFs controlled by dynamic cell type-specific super-enhancers that become repressed in both lineages. AHR and GLIS1 control differentiation-induced genes and their overexpression can inhibit the lineage commitment of the multipotent bone marrow-derived ST2 cells..\" \r\n\r\n...taken from D G\u00e9rard, et.al. Nucleic Acids Research, Volume 47, Issue 3, 20 February 2019, Pages 1141\u20131163, https://doi.org/10.1093/nar/gky1240",
+        "has_institutional_ethics_approval": true,
+        "has_national_ethics_approval": true,
+        "institutional_ethics_approval_notes": "The umbrella project LUX-Epigen has an institutional and a national ethics approval.",
+        "national_ethics_approval_notes": "The umbrella project LUX-Epigen has an institutional and a national ethics approval.",
+        "start_date": "2020-03-02",
+        "end_date": "2021-04-30",
         "contacts": [
             {
                 "first_name": "Alice",
                 "last_name": "White",
+                "email": "alice.white@uni.lu",
                 "role": "Principal_Investigator",
                 "institution": "ELU_I_77"
             },
             {
                 "first_name": "John",
-                "last_name": "Doe",
+                "last_name": "Black",
+                "email": "john.black@uni.lu",
                 "role": "Principal_Investigator",
                 "institution": "ELU_I_77"
             },
             {
-                "first_name": "Jane",
-                "last_name": "Doe",
+                "first_name": "Roy",
+                "last_name": "Blue",
+                "email": "roy.blue@uni.lu",
                 "role": "Researcher",
                 "institution": "ELU_I_77"
             }
         ],
-        "has_institutional_ethics_approval": false,
-        "has_national_ethics_approval": false,
-        "national_ethics_approval_notes": "Not needed as it is a software development project. No data collection.",
         "publications": [
             {
-                "citation_string": "Provenance-enabled stewardship of human data in the GDPR era. Pinar Alper, Regina Becker, Venkata Satagopam, Christophe Trefois, Valentin Groues, Jacek Lebioda, Yohan Jarosz.  ProvenanceWeek 2018 IPAW Proceedings, London, July 9-10 2018."
+                "citation": "Gérard D, Schmidt F, Ginolhac A, Schmitz M, Halder R, Ebert P, Schulz MH,\nSauter T, Sinkkonen L. Temporal enhancer profiling of parallel lineages\nidentifies AHR and GLIS1 as regulators of mesenchymal multipotency. Nucleic Acids\nRes. 2019 Feb 20;47(3):1141-1163. PubMed PMID:\n30544251; PubMed Central PMCID: PMC6380961.",
+                "doi": "doi: 10.1093/nar/gky1240"
+            }
+        ]
+    },
+    {
+        "acronym": "LUX-Epigen",
+        "name": "LUX-Epigen Research Programme - Phase1",
+        "description": "This is an imaginary research project created to demonstrate features of DAISY. LUX-Epigen is created as an umbrella project, which holds an ethics approval and contract among the participant institutes. There can be several PhD and Postdoc sub-projects cover under the umbrella  of Lux-Epigen.",
+        "has_institutional_ethics_approval": true,
+        "has_national_ethics_approval": true,
+        "institutional_ethics_approval_notes": "See attached Uni-LU ERP approval. This is a dummy document for the imaginary  LUX-Epigen programme.",
+        "national_ethics_approval_notes": "See attached CNER approval. This is a dummy document for the imaginary  LUX-Epigen programme.",
+        "start_date": "2017-04-03",
+        "end_date": "2022-04-25",
+        "contacts": [
+            {
+                "first_name": "Alice",
+                "last_name": "White",
+                "email": "alice.white@uni.lu",
+                "role": "Principal_Investigator",
+                "institution": "ELU_I_77"
             },
             {
-                "citation_string": "ELIXIR-Luxembourg: providing sustainability to clinical and translational medicine data for research. Venkata Satagopam, Pinar Alper, Regina Becker, Dietlind Gerloff, Wei Gu, Roland Krause, Jacek Lebioda, Noua Toukourou, Christophe Trefois, Reinhard Schneider. Poster presentations at Intelligent Systems for Molecular Biology 2018, ISMB, Chicago, July 6-10 2018."
+                "first_name": "John",
+                "last_name": "Black",
+                "email": "john.black@uni.lu",
+                "role": "Principal_Investigator",
+                "institution": "ELU_I_77"
+            },
+            {
+                "first_name": "Igor",
+                "last_name": "Green",
+                "email": "igor.green@chl.lu",
+                "role": "Principal_Investigator",
+                "institution": "ELU_I_9"
             }
         ]
     }
